tested to ensure that routes and seeds still work; changed necessary imports to esnure routes and seeds work and moved .env to server folder
