+++ server/routes/api/category-routes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const {Category, Product} = require('../../../models');
+const {Category, Product} = require('../../models');
 
 router.get('/', async (req, res) => {
     try {
