+++ server/routes/api/product-routes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { Product, Category, Tag, ProductTag } = require('../../../models');
+const { Product, Category, Tag, ProductTag } = require('../../models');
 
 router.get('/', async (req, res) => {
     try {
