+++ server/seeds/product-tag-seeds.js
@@ -1,4 +1,4 @@
-const { ProductTag } = require('../../models');
+const { ProductTag } = require('../models');
 
 const productTagData = [
   {
